add some Windows feature

+++ src/flat_hash_map/new_faster_hashtable.hpp
@@ -1,13 +1,17 @@
 #pragma once
 
 #include <algorithm>
+#include <cmath>
 #include <cstddef>
 #include <cstdint>
 #include <iterator>
-#include <math.h>
 #include <memory>
+#include <stdexcept>
 #include <type_traits>
 #include <utility>
+
+#include <iostream>
+#include "memory_pool.hpp"
 
 #ifdef _MSC_VER
 #define DDAOF_NOINLINE(...) __declspec(noinline) __VA_ARGS__
@@ -763,7 +767,7 @@ private:
     float _max_load_factor = 0.5f;
     size_t _num_elements = 0;
 
-    static size_t compute_max_lookups(size_t num_buckets) {
+    static int8_t compute_max_lookups(size_t num_buckets) {
         int8_t desired = log2(num_buckets);
         return std::max(desired, ddaof::min_lookups);
     }
@@ -1240,6 +1244,7 @@ private:
 };
 
 template<typename K, typename V, typename H = std::hash<K>, typename E = std::equal_to<K>, typename A = std::allocator<std::pair<K, V>> >
+// template<typename K, typename V, typename H = std::hash<K>, typename E = std::equal_to<K>, typename A = ddaof::AmnesiaAllocator<std::pair<K, V>> >
 class flat_hash_map : public ddaof::faster_hashtable<
             std::pair<K, V>, K,
             ddaof::key_or_value_hasher<K, std::pair<K, V>, H>, H,
